+++ src/dev/ipsych0/myrinnia/hpoverlay/HPOverlay.java
@@ -200,9 +200,9 @@ public class HPOverlay implements Serializable {
 
         // XP bar
         g.setColor(xpColor);
-        g.fillRoundRect(xpBar.x + 2, xpBar.y + 1, (int) (xpBar.width * Handler.get().getSkill(SkillsList.COMBAT).getExperience() / Handler.get().getSkill(SkillsList.COMBAT).getNextLevelXp()) - 2, xpBar.height - 4, 2, 4);
+        g.fillRoundRect(xpBar.x + 2, xpBar.y + 1, xpBar.width * Handler.get().getSkill(SkillsList.COMBAT).getExperience() / Handler.get().getSkill(SkillsList.COMBAT).getNextLevelXp() - 2, xpBar.height - 4, 2, 4);
         g.setColor(xpColorOutline);
-        g.drawRoundRect(xpBar.x + 2, xpBar.y + 1, (int) (xpBar.width * Handler.get().getSkill(SkillsList.COMBAT).getExperience() / Handler.get().getSkill(SkillsList.COMBAT).getNextLevelXp()) - 2, xpBar.height - 4, 2, 4);
+        g.drawRoundRect(xpBar.x + 2, xpBar.y + 1, xpBar.width * Handler.get().getSkill(SkillsList.COMBAT).getExperience() / Handler.get().getSkill(SkillsList.COMBAT).getNextLevelXp() - 2, xpBar.height - 4, 2, 4);
 
         Text.drawString(g, "Combat level: " + Integer.toString(Handler.get().getSkillsUI().getSkill(SkillsList.COMBAT).getLevel()),
                 combatBar.x + combatBar.width / 2, combatBar.y + combatBar.height / 2, true, Color.YELLOW, Assets.font14);
